Rename file for case consistency
